+++ src/engine/sphere.h
@@ -7,7 +7,7 @@ struct ray;
 
 struct sphere {
     struct vector position;
-    float radius;
+    double radius;
 };
 
 optional_vector sphere_ray_intersection(void* sphere, struct ray* ray);
